fix: corregir organizacion de carpetas
