--- Java/src/main/java/org/sisuns/bucles_for/Ejercicio_04.java
@@ -1,4 +1,0 @@
-package org.sisuns.bucles_for;
-
-public class Ejercicio_04 {
-}
