+++ Java/src/main/java/org/sisuns/bucles_for/avanzado/Ejercicio_02.java
@@ -1,0 +1,4 @@
+package org.sisuns.bucles_for.avanzado;
+
+public class Ejercicio_02 {
+}
