+++ Java/src/main/java/org/sisuns/bucles_for/intermedio/Ejercicio_01.java
@@ -1,0 +1,4 @@
+package org.sisuns.bucles_for.intermedio;
+
+public class Ejercicio_01 {
+}
