+++ Java/src/main/java/org/sisuns/bucles_for/intermedio/Ejercicio_02.java
@@ -1,0 +1,4 @@
+package org.sisuns.bucles_for.intermedio;
+
+public class Ejercicio_02 {
+}
