--- Java/src/main/java/org/sisuns/bucles_while/Ejercicio_03.java
@@ -1,4 +1,0 @@
-package org.sisuns.bucles_while;
-
-public class Ejercicio_03 {
-}
