+++ Java/src/main/java/org/sisuns/bucles_while/basico/Ejercicio_05.java
@@ -1,0 +1,4 @@
+package org.sisuns.bucles_while.basico;
+
+public class Ejercicio_05 {
+}
