+++ Java/src/main/java/org/sisuns/bucles_while/intermedio/Ejercicio_01.java
@@ -1,0 +1,4 @@
+package org.sisuns.bucles_while.intermedio;
+
+public class Ejercicio_01 {
+}
