+++ Java/src/main/java/org/sisuns/bucles_while/intermedio/Ejercicio_04.java
@@ -1,0 +1,4 @@
+package org.sisuns.bucles_while.intermedio;
+
+public class Ejercicio_04 {
+}
