--- Java/src/main/java/org/sisuns/condicionales/Ejercicio_02.java
@@ -1,4 +1,0 @@
-package org.sisuns.condicionales;
-
-public class Ejercicio_02 {
-}
