+++ Java/src/main/java/org/sisuns/condicionales/avanzado/Ejercicio_01.java
@@ -1,0 +1,4 @@
+package org.sisuns.condicionales.avanzado;
+
+public class Ejercicio_01 {
+}
