+++ Java/src/main/java/org/sisuns/condicionales/avanzado/Ejercicio_05.java
@@ -1,0 +1,4 @@
+package org.sisuns.condicionales.avanzado;
+
+public class Ejercicio_05 {
+}
