+++ Java/src/main/java/org/sisuns/condicionales/basico/Ejercicio_02.java
@@ -1,0 +1,4 @@
+package org.sisuns.condicionales.basico;
+
+public class Ejercicio_02 {
+}
