+++ Java/src/main/java/org/sisuns/condicionales/intermedio/Ejercicio_04.java
@@ -1,0 +1,4 @@
+package org.sisuns.condicionales.intermedio;
+
+public class Ejercicio_04 {
+}
